requirements.txt
# =====================================================
# DAVID CYBER INTELLIGENCE SYSTEM - Requirements
# Supports: Windows | macOS | Linux
# =====================================================

# --- Core API & Server ---
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
python-dotenv>=1.0.0
httpx>=0.27.0
requests>=2.31.0

# --- LLM Brain ---
ctransformers>=0.2.27
transformers>=4.40.0
tokenizers>=0.19.0

# --- ML / AI ---
torch>=2.2.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0

# --- Malware Engine ---
yara-python>=4.5.0
pefile>=2023.2.7
capstone>=5.0.1

# --- Network Engine ---
scapy>=2.5.0
pyshark>=0.6

# --- OSINT Engine ---
shodan>=1.31.0
# theHarvester -> install separately: pip install theHarvester
# spiderfoot -> install separately: pip install spiderfoot

# --- NER / NLP ---
spacy>=3.7.0
# CyNER -> install from GitHub: pip install git+https://github.com/aiforsec/CyNER

# --- Satellite Tracking ---
skyfield>=1.48
sgp4>=2.22

# --- Geospatial ---
geopandas>=0.14.0
shapely>=2.0.0
folium>=0.16.0
geoip2>=4.8.0

# --- Database ---
psycopg2-binary>=2.9.9
elasticsearch>=8.13.0
neo4j>=5.19.0
sqlalchemy>=2.0.29

# --- Threat Intel ---
pymisp>=2.4.191
# OpenCTI Python client
pycti>=6.1.5

# --- Visualization ---
plotly>=5.20.0
matplotlib>=3.8.0
leaflet  # via frontend

# --- Utilities ---
rich>=13.7.0
loguru>=0.7.2
aiofiles>=23.2.1
tqdm>=4.66.0
click>=8.1.7
cryptography>=42.0.0
